♻️ separate core and example
