+++ packages/atom.io/__tests__/experimental/immortal/allocate.test.ts
@@ -1,7 +1,19 @@
 import type { Logger } from "atom.io"
 import { clearStore, IMPLICIT } from "atom.io/internal"
 
-import { allocateIntoStore } from "~/packages/atom.io/src/allocate"
+import type {
+	Above,
+	Below,
+	Claim,
+	Hierarchy,
+	TypedKey,
+	Vassal,
+} from "~/packages/atom.io/src/allocate"
+import {
+	allocateIntoStore,
+	createAllocator,
+	T$,
+} from "~/packages/atom.io/src/allocate"
 
 const LOG_LEVELS = [null, `error`, `warn`, `info`] as const
 const CHOOSE = 2
@@ -17,13 +29,67 @@ beforeEach(() => {
 	vitest.spyOn(logger, `warn`)
 	vitest.spyOn(logger, `info`)
 })
-describe(`moleculeFamily`, () => {
-	test(`exclusive molecule hierarchy`, () => {
-		allocateIntoStore(IMPLICIT.STORE, `root`, `a`)
-		allocateIntoStore(IMPLICIT.STORE, `root`, `b`)
-		allocateIntoStore(IMPLICIT.STORE, `root`, `c`)
-		allocateIntoStore(IMPLICIT.STORE, `root`, `d`)
-		allocateIntoStore(IMPLICIT.STORE, `root`, `e`)
-		allocateIntoStore(IMPLICIT.STORE, `root`, `f`)
+describe(`allocate`, () => {
+	test(`the Hierarchy + allocate + claim pattern`, () => {
+		type myClaim = Claim<Hierarchy, TypedKey, []>
+
+		type GameKey = [`game`, string]
+		type UserKey = [`user`, string]
+		type PlayerKey = [[T$, `player`], GameKey, UserKey]
+		type ItemKey = [`item`, string]
+
+		type GameHierarchy = Hierarchy<
+			[
+				{
+					above: `root`
+					below: [GameKey, UserKey]
+				},
+				{
+					above: [GameKey, UserKey]
+					below: [PlayerKey]
+				},
+				{
+					above: [GameKey]
+					below: [ItemKey]
+				},
+				{
+					above: [PlayerKey]
+					below: [ItemKey]
+				},
+			]
+		>
+
+		type GameVassal = Vassal<GameHierarchy>
+
+		type AboveGame = Above<GameKey, GameHierarchy>
+		type AboveUser = Above<UserKey, GameHierarchy>
+		type AbovePlayer = Above<PlayerKey, GameHierarchy>
+		type AboveItem = Above<ItemKey, GameHierarchy>
+
+		type BelowGame = Below<[GameKey], GameHierarchy>
+		type BelowUser = Below<[UserKey], GameHierarchy>
+		type BelowGameUser = Below<[GameKey, UserKey], GameHierarchy>
+		type BelowPlayer = Below<[PlayerKey], GameHierarchy>
+		type BelowItem = Below<[ItemKey], GameHierarchy>
+
+		const gameKey = [`game`, `xxx`] satisfies GameKey
+		const userKey = [`user`, `yyy`] satisfies UserKey
+		const playerKey = [[T$, `player`], gameKey, userKey] satisfies PlayerKey
+		const gameClaim0 = allocateIntoStore(IMPLICIT.STORE, `root`, gameKey)
+		const userClaim0 = allocateIntoStore(IMPLICIT.STORE, `root`, userKey)
+		const playerClaim0 = allocateIntoStore(
+			IMPLICIT.STORE,
+			[gameClaim0, userClaim0],
+			playerKey,
+		)
+
+		const gameAllocator = createAllocator<GameHierarchy>(IMPLICIT.STORE)
+
+		const gameClaim = gameAllocator(`root`, gameKey)
+		const userClaim = gameAllocator(`root`, userKey)
+		const playerClaim = gameAllocator([gameClaim, userClaim], playerKey)
+
+		const itemKey = [`item`, `xxx`] as [`item`, string]
+		const itemClaim = allocateIntoStore(IMPLICIT.STORE, [playerClaim], itemKey)
 	})
 })
